change autoblaster to blastercomponent, and add autofirecomponent
