+++ assets/data/enemies.ron
@@ -50,7 +50,7 @@
             forward: true,
             animation_type: PingPong,
         ),
-        autoblaster_component: Some((
+        blaster_component: Some((
             count: 1,
             allied: false,
             shot_velocity: [0.0, -80.0],
@@ -62,6 +62,8 @@
             crit_chance: 0.0,
             size_multiplier: 1.0,
             spacing: 7.0,
+        )),
+        autofire_component: Some((
             fire_period: 3.0,
             fire_timer: 1.0,
         )),
@@ -143,7 +145,7 @@
             forward: true,
             animation_type: PingPong,
         ),
-        autoblaster_component: Some((
+        blaster_component: Some((
             count: 1,
             allied: false,
             shot_velocity: [0.0, -70.0],
@@ -155,6 +157,8 @@
             crit_chance: 0.0,
             size_multiplier: 1.0,
             spacing: 7.0,
+        )),
+        autofire_component: Some((
             fire_period: 3.0,
             fire_timer: 1.0,
         )),
